backend/core/requirements.txt
fastapi
uvicorn
requests
qdrant_client
sentence_transformers
langchain_text_splitters
pydantic
openai==0.28
python-multipart 
python-dotenv   
opencv-python-headless
PyMuPDF==1.19.0
opencv-python
pdfplumber
pytesseract
minio
psycopg2-binary
bcrypt
motor 
python-jose[cryptography]
pdf2image==1.16.0
pillow==8.4.0